+++ app/src/main/java/org/sopt/sample/di/DataSourceModule.kt
@@ -8,10 +8,10 @@ import dagger.hilt.components.SingletonComponent
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
-import org.sopt.sample.application.AppInterceptor
-import org.sopt.sample.application.AuthService
-import org.sopt.sample.application.ReqresApi
-import org.sopt.sample.application.Util.Constant
+import org.sopt.sample.data.Util.AppInterceptor
+import org.sopt.sample.data.datasource.AuthService
+import org.sopt.sample.data.datasource.ReqresApi
+import org.sopt.sample.data.Util.Constant
 import retrofit2.Retrofit
 import javax.inject.Singleton
